fix: reducing complexity
