+++ apps/artemis-web/pages/styles.css
@@ -97,10 +97,6 @@ footer > .MuiContainer-root {
   max-width: 85% !important;
 }
 
-.card {
-  border: 0px;
-}
-
 .caret {
   top: -25px;
   left: 50%;
